Switch away from reusing the same cursor
Fix #91

This introduces several bugs where one query can affect the next one.

In the case exercised by the modified test, `expire_lists` wasn't
functioning properly. It queried all lists with SELECT, but when it
found a list to expire, the DELETE-query overrode the still "active"
SELECT-query and "cleared" the iterator.

This is comparable to modifying a dictionary while iterating through it,
except sqlite appears not to give us any warning about that.

There are two ways to fix this problem:

- Drain the iterator using `list(result)` or `result.fetchall()`
- Just use a different cursor for each query

I didn't know that the latter was an option. When I wrote the original
code that used a global cursor everywhere, I encountered very weird
testfailures when explicitly instantiating a cursor for `add_file`,
where the database lock timed out. But I can't reproduce those failures
anymore.