Make Formatter API cleaner
Make the TZ configurable: we were already hacking to change it, so make
it part of its actual API. Note that this has no user-facing changes,
only the internal API (mostly for tests).

Also default formats to the DefaultFormatter. This makes subclassing
simpler, but provides the same API benefits as the above change.