Add test to for uid/filename colition on save
Due to how our save logic is implemented, I suspected that this
particular scenario would have a silent failure when attempting to save.

Fortunately, we pre-define a Todo's filename on-initialization, so it's
not a bug (merely by coincidence).

Let's keep the test around just in case we introduce a regression like
this in future.