Read displayname files in calendars directories
Don't stop reading configuration file when a `path = ` is encountered.
Build an array instead of all values of `path = ` in configuration file.
Use every element in that array and expand any `*` in it and don't rely
on the existence of a `displayname` file in the expanded directories.
In case a directory without a `displayname` file in it is encountered,
we revert to suggest the name of the directory itself.