Use a fluid layout for TODOs
Stop drawing a table as done previously, and use a fluid layout. This
reduces internal whitespace, which makes the ids for todos easier to
recognise in may scenarios.

Also makes the description and location blocks easier to copy-paste
verbatim without additional indentation.

Due to dropping the table rendering, this should slightly improve
performance, though that's not a main goal here.