fix filtering for non-ASCII categories
To make category filtering case insensitive, todoman converts the
category provided on the commandline to uppercase, uses the sqlite
`upper` function to convert the cache entries and then compares the
converted strings.

However the sqlite `upper` function only works for ASCII strings causing
problems when the category is for instance `ý`, which sqlite translates
to uppercase as `ý` while the correct version is `Ý`, causing the
filtering for this category not to return any tasks.

Add a new column (`category_upper`) to the `categories` table with the
uppercase version of the category name converted by Python (which
performs the conversion correctly) and use this for filtering while
keeping the original version in the `category` column as before.

Add a corresponding test.