feat!: add support for `RELATED-TO` property and `RELTYPE` parameter
Taken from RFC 5545, section `3.8.4.5`. If present in a file,
these values are now read and stored in the cache. Nothing in the UI
has been changed. Tests are modified to reflect the change.

BREAKING CHANGE: The cache has to be deleted manually and recreated if
the cache was previously created using a version of the code,
that is older than this commit.